created components for list in program and favs, now more code is shared between, and more dry
