+++ front/src/components/Pages/Favs.vue
@@ -2,25 +2,32 @@
   <div class="container">
     <div class="tab-container col s12 m7">
 
-            <div v-for="programItem in programItems" :key="programItem.tabId"  v-if="activeProgramTab == programItem.tabId" class="col s12 tabsingle">
-                <div v-for="programHours in programItem.hours">
-                    <div class="collection-item-time" v-if="shouldShowHourLabel(programHours.activities)"><span>{{programHours.hour}}</span></div>
-                    
-                    <ul class="collection">
-                        <li v-for="programActivity in programHours.activities" 
-                            v-if="favItems.indexOf(programActivity.id) > -1" 
-                            :key="programActivity.id" 
-                            class="collection-item" 
-                            :class="'collection-item-'+programActivity.type">
+        <div 
+            v-for="programItem in programItems" 
+            v-if="activeProgramTab == programItem.tabId" 
+            :key="programItem.tabId"  
+            class="col s12 tabsingle">
+            
+            <div 
+                v-for="programHours in programItem.hours" 
+                :key="programHours.hour">
 
-                            <span class="title">{{programActivity.title}}</span>
-                            <p>Das {{programActivity.hourStart}} às {{programActivity.hourEnd}}</p>
+                <CollectionTime 
+                    v-if="shouldShowHourLabel(programHours.activities)" 
+                    :hour="programHours.hour">
+                </CollectionTime>
 
-                            <a v-on:click="unfavActivity(programActivity.id)" class="secondary-content orange-text text-lighten-1"><i class="material-icons">star</i></a>
-                        </li>
-                    </ul>
-                </div>
+                <ul class="collection">
+                    <CollectionItem  
+                        v-for="programActivity in programHours.activities" 
+                        v-if="favItems.indexOf(programActivity.id) > -1" 
+                        :programActivity="programActivity" 
+                        :key="programActivity.id" >
+                    </CollectionItem>
+                </ul>
+
             </div>
+        </div>
 
     </div>
 
@@ -29,16 +36,17 @@
 
 
 <script>
+import CollectionItem from "../Shared/CollectionItem"
+import CollectionTime from "../Shared/CollectionTime"
+
 export default {
     name: 'program',
+    components: {CollectionItem, CollectionTime},
     data () {
         return {
         }
     },
     methods: {
-        unfavActivity: function(activityId) {
-            this.$store.dispatch('unsaveFav', activityId);
-        },
         shouldShowHourLabel: function(hourActivities) {
             let shouldShow = false
 
@@ -105,59 +113,6 @@ export default {
         }
     }
 
-    .collection-item {
-        padding: 16px 22px 13px;
-        border-left: 5px solid #333;
-        border-bottom: none;
-        position: relative;
-
-        .secondary-content {
-            position: absolute;
-            top: 23px;
-            right: 16px;
-            cursor: pointer;
-        }
-
-        span.title {
-            font-size: 14px;
-            font-weight: 600;
-            text-transform: uppercase;
-        }
-
-        p {
-            margin: 0;
-        }
-
-        &.collection-item-party {
-            span.title {color:#7F6D9A}
-            border-color: #7F6D9A;
-        }
-        &.collection-item-cone {
-            span.title {color:#46BBE6}
-            border-color: #46BBE6;
-        }
-        &.collection-item-content {
-            span.title {color:#E67A9F}
-            border-color: #E67A9F;
-        }
-        &.collection-item-morning {
-            span.title {color:#82A75E}
-            border-color: #82A75E;
-        }
-        &.collection-item-lunch {
-            span.title {color:#E19255}
-            border-color: #E19255;
-        }
-        &.collection-item-other {
-            span.title {color:#61BBB7}
-            border-color: #61BBB7;
-        }
-        &.collection-item-special {
-            span.title {color:#AEAFAE}
-            border-color: #AEAFAE;
-        }
-    }
-
     .collection-empty {
         height: 100%;
         display: flex;
@@ -170,43 +125,6 @@ export default {
             color: #bbb;
             padding-bottom: 25vh;
         }
-    }
-
-    .collection-item-time {
-        border-left: none;
-        background: #f3f3f3;
-        font-size: 14px;
-        color: #666;
-        padding: 8px 0px 2px 25px;
-
-        @media only screen and (min-width: 660px) {
-            padding: 18px 0px 2px 32px;
-            background: none; 
-            
-            span {
-                color: #666;
-                font-size: 16px;
-                font-weight: 600;
-            }
-        }
-    }
-
-    .slidetab-enter-active, .slidetab-leave-active {
-        transition: all .5s
-    }
-    .slidetab-enter {
-        top: 60px;
-        opacity: 0;
-    }
-
-    .slidetab-leave-to {
-        top: -60px;
-        opacity: 0;
-    }
-
-    .slidetab-enter-to, .slidetab-leave {
-        top: 0px;
-        opacity: 1;
     }
 
 </style>
